requirements.txt
cognite-sdk==0.12.12
